requirements.txt
altgraph==0.17.5
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
et_xmlfile==2.0.0
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
httplib2==0.31.0
idna==3.11
numpy==2.4.0
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pefile==2024.8.26
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodomex==3.23.0
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
pyodbc==5.3.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
pyzipper==0.3.6
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
xlsxwriter==3.2.9
